refactor: 시도 횟수 입력 클래스에서 검증하는 클래스 분리
