+++ src/main/java/racingcar/game/UserInputGameRound.java
@@ -16,38 +16,7 @@ public class UserInputGameRound {
     }
 
     public static int processInput(String round) {
-        checkIsBlank(round);
-        int gameRound = checkIsNumber(round);
-        checkIsZero(gameRound);
-
-        return gameRound;
-    }
-
-    private static int checkIsNumber(String gameRound) {
-        for (char digit : gameRound.toCharArray()) {
-            validateDigit(digit);
-        }
-
-        return Integer.parseInt(gameRound);
-    }
-
-    private static void validateDigit(char digit) {
-        int num = digit - '0';
-
-        if (num < Constant.MIN_VALUE || num > Constant.MAX_VALUE) {
-            throw new IllegalArgumentException("0부터 9까지의 숫자를 입력해주세요.");
-        }
-    }
-
-    private static void checkIsZero(int num){
-        if(num <= Constant.MIN_VALUE){
-            throw new IllegalArgumentException("시도 횟수 0 이하입니다.");
-        }
-    }
-
-    private static void checkIsBlank(String gameRound){
-        if (gameRound.isEmpty()) {
-            throw new IllegalArgumentException("입력값이 공백입니다.");
-        }
+        ValidateGame.validator(round);
+        return Integer.parseInt(round);
     }
 }
